requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
#blinker==1.8.2
blinker==1.9.0
catalogue==2.0.10
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
confection==0.1.5
cryptography==43.0.0
databases==0.9.0
exceptiongroup==1.2.2
fastapi~=0.115.11
Flask~=3.1.0
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==8.2.0
importlib_resources==6.4.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ollama==0.3.1
packaging==24.1
pluggy==1.5.0
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
PyMySQL==1.1.1
pytest==8.3.2
python-dotenv==1.0.1
requests==2.32.3
rich==13.7.1
shellingham==1.5.4
smart-open==7.0.4
sniffio==1.3.1
SQLAlchemy==2.0.32
srsly==2.4.8
starlette~=0.46.1
tomli==2.0.1
typer==0.12.3
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.5
wasabi==1.1.3
weasel==0.4.1
Werkzeug~=3.1.3
wrapt==1.16.0
zipp==3.19.2

pillow~=11.1.0
setuptools~=75.1.0
protobuf~=5.29.3
openai~=1.60.1
pandas~=2.2.3

projectdavid